Split mutate.ts; relocate terminal settings migration
D2-3: split mutate.ts into mutate-steps (## Steps), mutate-status
(Status + ## Timeline), write-config (note write + config
canonicalisation), over a shared mutate-shared text toolkit; mutate.ts
becomes a re-export barrel so existing imports resolve unchanged.
D2-4: move the one-shot migrateTerminalFromConfigIfNeeded out of the PTY
session manager into settings-migrations.ts.
