+++ src/main/index.ts
@@ -10,7 +10,8 @@ import { DEFAULT_LAYOUT, readSettings } from './settings';
 import { setWatchedConception } from './watcher';
 import { setScheduledConception } from './task-scheduler';
 import { disposeRepoWatchers } from './repo-watchers';
-import { killAll, migrateTerminalFromConfigIfNeeded } from './terminals';
+import { killAll } from './terminals';
+import { migrateTerminalFromConfigIfNeeded } from './settings-migrations';
 import { loginPath } from './shell-env';
 import { runLogJanitor } from './terminal-logger-janitor';
 import { sealOrphanLogs } from './seal-orphan-logs';
